refactor: Resolve circular dependencies
